feat: add existing code
adds in existing solution written for use with training data. content has been slightly modified to remove hard coded references to individual submission data in the training case, for privacy purposes. that content can still be used for the actual run, but is now specified as a parameterized set of override files that are not included in the repo.