tests: add userspace option to check results versus known matches
testing is hard in this model, but here's a swing at it: once you have a curated training set, you can specify it as a filename at runtime and any queries that match the training set (exactly) will be checked for consistency versus the curated set. works well so far? can't post the training data to github but at least I can run it locally.

re #2