fix: handle situations where both entry formats match
initially, entries were supposed to be either "TITLE - AUTHOR" or "TITLE by AUTHOR". and that's great! except finally at long last, someone with a hyphenated last name showed up to ruin my morning. so new behavior for handling format collisions is to prioritize the "(.*) by (.*)" match. this is imperfect ,and requires some overrides; but then, that's why I added the override functionality to begin with lol.